renamed definitions to parsed-nodes
